requirements.txt
# Dependências do Projeto
# Instale com: pip install -r requirements.txt

# Kafka
kafka-python==2.0.2

# Utilitários para JSON e data
python-dateutil==2.8.2

# Logging melhorado (opcional)
colorlog==6.7.0

# Monitoramento (opcional)
prometheus-client==0.17.1

# Para desenvolvimento
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Para containers
docker==6.1.3

# Para Kubernetes (opcional)
kubernetes==27.2.0
